requirements.txt
Django==4.0.6
django-browser-reload==1.6.0
PyMySQL==1.1.0
xhtml2pdf==0.2.11
django_bootstrap5==23.3
python-decouple==3.8
stripe==7.1.0
djangorestframework==3.13.1